N_i needs to be of type
leads to errors in .synapses when taking shape
otherwise

+++ examples/zheng2013/param_noinput.py
@@ -7,7 +7,7 @@ utils.backup(__file__)
 from common.defaults import *
 
 c.N_e = 200
-c.N_i = np.floor(0.2*c.N_e)
+c.N_i = int(np.floor(0.2*c.N_e))
 c.N_u = 0 #noinput is the point of this experiment
 c.N = c.N_e + c.N_i
